ci: Use side-effect-free version of ngx-markdown config/factory (to reduce noise in tests)
