+++ src/app/components/offer/offer.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { provideRouter } from '@angular/router';
 import mockOffer from 'src/app/mocks/offer.mock';
 import { LoggingService } from 'src/app/services/logging.service';
-import { ngxMarkdownModuleFactory } from 'src/main';
+import { ngxMarkdownModuleFactory } from 'src/app/shared/ngx-markdown.factory';
 
 import { OfferComponent } from './offer.component';
 
@@ -12,8 +12,9 @@ describe('OfferComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      imports: [provideRouter([]), ngxMarkdownModuleFactory()],
+      imports: [ngxMarkdownModuleFactory()],
       providers: [
+        provideRouter([]),
         {
           provide: LoggingService,
         },
